# ⚡️ Speed up InMemoryCache.evict_cache by 21%